Improve speed of day06b solution
The solution gets the bounding box of all points, expands it by 10000 (the maximum allowed sum-of-distances), then iterates over every integer coordinate in it. I realised that expanding by the maximum distance is too far, we only need to expand by 10000 / N, where N is the number of points. This is because the metric we care about is the sum of the distances to every point, which goes up by N rather than 1 for each X and Y step outside of the bounding box. This gives us a much smaller area to iterate over, giving a massive speed up.

A second smaller optimisation was that we only care about points whose manhattan distance is within 10000/N of the original bounds. This is a diamond-shaped region, so there are a fair few coordinates near the corners of our expanded bounding box that still can't possibly be within the range we care about. We can use the manhattan distance between each coordinate and the nearest point on or in the (non-expanded) bounding box as a quick rejection test, to avoid having to iterate over all points when their sum can't possibly be less than 10000.

I had another idea, which was that we could calculate the sum of the distances along each axis separately and store the results in a pair of arrays. Then we just need to find the pairs of array indexes where the array values add up to less than 10000. I haven't bothered to implement that yet though, because what I have is fast enough now (it's gone from 2.14 seconds to 23 milliseconds).