Improve performance of my day22b solution
Instead of caching geologic index, cache the terrain type instead so we don't have to recalculate it every time.

Cache an area of terrain with width and height both equal to the depth of the cave system. This will be large enough that my solution never needs to go outside it, so I can get rid of the std::map holding for outside of the cache.

Instead of using a set to keep track of the visited state for each cell, use an array with the same dimensions as the terrain cache. This array holds a bitset for each cell which says which of the 3 tools we've reached the cell with. This is a *lot* more efficient to get and set than a std::map.

Check whether we've already visited a particular cell with a particular tool *before* adding it to the work queue. Adding an item to the work queue is comparatively expensive, so this saves a fair bit of time.

When switching tools, note that there's only *one* other possibility, not two. Only queue up work for that alternate (and only if we haven't already visited our current cell with that tool). I also noticed that the alternate tool for any given terrain type can be calculated with a simple formula, no lookups or conditionals required.

The runtime for day22b has dropped from 2.3 seconds to 0.6 seconds with these changes.