More performance improvements for day22b
Limit the cached region to double the target coords and don't queue up work for cells outside this region. This may not be big enough for all possible cases, but is sufficient for my input. This gives an *enormous* speedup.

Prioritise work by distance from the target as well as by time.

Additional smaller improvements:
- Pack the terrain cache & visited flags into a single byte per cell.
- Use smaller member data types for ivec2 and Work (int16_t and uint16_t instead of int).

Runtime has dropped from 0.5 seconds to 0.013 seconds!